fix: ranking score fix

+++ src/main/java/com/example/demo/Ranking/DTO/RankingResponseDTO.java
@@ -11,5 +11,5 @@ public class RankingResponseDTO {
 
     private String userEmail;
 
-    private Long score;
+    private Double score;
 }
